mv and rename TldrawCanvasLabelDialog
