+++ src/components/tldraw/DiscourseNode.tsx
@@ -21,7 +21,7 @@ import createDiscourseNode from "../../utils/createDiscourseNode";
 import getDiscourseContextResults from "../../utils/getDiscourseContextResults";
 import { measureCanvasNodeText } from "../../utils/measureCanvasNodeText";
 import { COLOR_ARRAY, DEFAULT_STYLE_PROPS, discourseContext } from "./Tldraw";
-import LabelDialog from "../TldrawCanvasLabelDialog";
+import LabelDialog from "./LabelDialog";
 import { isPageUid } from "../../utils/isPageUid";
 import { loadImage } from "../../utils/loadImage";
 import { DiscourseRelationShape } from "./DiscourseRelations";
